refactor: PropertyPanel into 5 files
